Pantalla de curso completada.

+++ SchoolProSite.DAL/Models/CourseDaoModel.cs
@@ -5,7 +5,7 @@ namespace SchoolProSite.DAL.Models
     {
         public int CourseId { get; set; }
         public string? Title { get; set; }
-        public decimal Credits { get; set; }
+        public int Credits { get; set; }
         public int DepartmentId { get; set; }
         public string? DepartmentName { get; set; }
         public DateTime CreationDate { get; set; }
